Rename AudObstruction to AudObstructionOcclusion and add doxygen
